// src/lib/types.ts
export type MessageReaction = {
  id: string;
  emoji: string;
  userId: string;
  userName: string;
  createdAt: string;
};